@@ -4,7 +4,7 @@
     "description": "Enum base class",
     "license": "MIT",
     "require": {
-        "php": "^7.1 || ^8.0"
+        "php": "^8.2"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5"
