@@ -4,45 +4,58 @@ declare(strict_types=1);
 
 namespace Iresults\Enum;
 
-use InvalidArgumentException;
 use Iresults\Enum\Exception\EnumOutOfRangeException;
 use Iresults\Enum\Exception\InvalidEnumArgumentException;
 use ReflectionClass;
 
 use function array_key_exists;
 use function array_search;
-use function call_user_func;
-use function get_class;
 use function gettype;
-use function is_array;
 use function is_null;
-use function is_object;
 use function is_scalar;
 use function is_string;
 use function sprintf;
 use function strtoupper;
 
-abstract class EnumFactory
+final class EnumFactory
 {
-    private static $registry = [];
+    /**
+     * @var array<string, EnumInterface<scalar|list<scalar>|null>>
+     */
+    private static array $registry = [];
 
-    private static $reflectionCache = [];
+    /**
+     * @var array<class-string, array<string,mixed>>
+     */
+    private static array $reflectionCache = [];
+
+    private function __construct()
+    {
+    }
 
     /**
      * Return the instance of the enum from the registry or create a new one
      *
-     * @param array|bool|float|int|string $valueOrName
+     * @template T of scalar|null|list<scalar>
+     *
+     * @param T|string                       $valueOrName
+     * @param class-string<EnumInterface<T>> $className
+     *
+     * @return EnumInterface<T>
      */
-    public static function makeInstance($valueOrName, string $className): EnumInterface
-    {
+    public static function makeInstance(
+        array|bool|float|int|string|null $valueOrName,
+        string $className,
+    ): EnumInterface {
         if (!static::isValidValueType($valueOrName)) {
-            throw new InvalidEnumArgumentException(sprintf('Type of value is not a valid constant type or name: "%s"', is_object($valueOrName) ? get_class($valueOrName) : gettype($valueOrName)));
+            throw new InvalidEnumArgumentException(sprintf('Type of value is not a valid constant type or name: "%s"', get_debug_type($valueOrName)));
         }
         InvalidEnumArgumentException::assertValidEnumClass($className);
 
         if (is_scalar($valueOrName)) {
             $argumentRegistryKey = 'argkey::' . $className . '::' . gettype($valueOrName) . '::' . $valueOrName;
             if (isset(self::$registry[$argumentRegistryKey])) {
+                // @phpstan-ignore return.type
                 return self::$registry[$argumentRegistryKey];
             }
         } else {
@@ -62,8 +75,8 @@ abstract class EnumFactory
         }
         $registryKey = 'namekey::' . $className . '::' . strtoupper($name);
         if (!isset(self::$registry[$registryKey])) {
-            $instance = call_user_func(
-                [$className, 'createInstance'],
+            // @phpstan-ignore staticMethod.notFound
+            $instance = $className::createInstance(
                 $value ?? $value = static::retrieveValueForName($name, $className),
                 $name
             );
@@ -76,10 +89,14 @@ abstract class EnumFactory
             return $instance;
         }
 
+        // @phpstan-ignore return.type
         return self::$registry[$registryKey];
     }
 
-    private static function getConstants(string $className)
+    /**
+     * @return array<string,mixed>
+     */
+    private static function getConstants(string $className): array
     {
         if (!isset(static::$reflectionCache[$className])) {
             $reflection = new ReflectionClass($className);
@@ -92,26 +109,29 @@ abstract class EnumFactory
     /**
      * Returns the if a constant with the given name exists
      */
-    private static function hasConstant(string $constantName, string $className): bool
-    {
-        if (!is_string($constantName)) {
-            throw new InvalidArgumentException('Expected argument "constantName" to be of type string');
-        }
-
-        return array_key_exists(strtoupper($constantName), static::getConstants($className));
+    private static function hasConstant(
+        string $constantName,
+        string $className,
+    ): bool {
+        return array_key_exists(
+            strtoupper($constantName),
+            static::getConstants($className)
+        );
     }
 
     /**
-     * Returns the constant name for the given value
+     * Return the constant name for the given value
      *
      * If the enum contains multiple constants with the given value the behaviour is undefined
      *
-     * @param int|float|string|array|bool $constantValue
+     * @param mixed[]|bool|float|int|string $constantValue
      *
      * @return string|bool Returns the name or FALSE if not found
      */
-    private static function getNameForValueOfClass($constantValue, string $className)
-    {
+    private static function getNameForValueOfClass(
+        array|bool|float|int|string|null $constantValue,
+        string $className,
+    ): bool|int|string {
         if (!static::isValidValueType($constantValue)) {
             return false;
         }
@@ -121,35 +141,38 @@ abstract class EnumFactory
             $descriptor = new EnumDescriptor();
         }
 
-        return array_search($constantValue, $descriptor->getValues($className), true);
+        return array_search(
+            $constantValue,
+            $descriptor->getValues($className),
+            true
+        );
     }
 
     /**
-     * Returns if the given value is a valid type for an enum in general
+     * Return if the given value is a valid type for an enum in general
      *
-     * @return bool
+     * @param mixed[]|bool|float|int|string|null $value
      */
-    private static function isValidValueType($value)
+    private static function isValidValueType(array|bool|float|int|string|null $value): bool
     {
         if (is_null($value) || is_scalar($value)) {
             return true;
         }
-        if (is_array($value)) {
-            // Loop through the elements of the array and return false if one of it is not a valid value type
-            foreach ($value as $element) {
-                if (!static::isValidValueType($element)) {
-                    return false;
-                }
-            }
 
-            return true;
+        // Loop through the elements of the array and return false if one of it is not a valid value type
+        foreach ($value as $element) {
+            if (!static::isValidValueType($element)) {
+                return false;
+            }
         }
 
-        return false;
+        return true;
     }
 
-    private static function retrieveValueForName(string $constantName, string $className)
-    {
+    private static function retrieveValueForName(
+        string $constantName,
+        string $className,
+    ): mixed {
         return static::getConstants($className)[strtoupper($constantName)] ?? null;
     }
 }
